+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 
 - Change mangrove-ts to a single package repo for mangrove.js.
+- Make gas estimation explicit to allow usage of ethers estimation.
 
 # 1.2.10